# Fix build performance regression