@anadale 2023-03-01
@martinothamar, I think that 3 level of deepness is more than enough for the vast majority of cases. The only thing, though, is that it may be worth to add a configuration, like for the namespace, so those, who "need to go deeper", can achieve what they want - even sacrificing the performance.